Scripts/requirements.txt
discord
asyncio
Pillow
pytesseract
requests